rename minting as create profile page
